cleanup and further work on chat

+++ src/wbs/framework/object/ObjectHelperBuilder.java
@@ -672,7 +672,7 @@ class ObjectHelperBuilder {
 			@Override
 			public
 			Record<?> find (
-					int id) {
+					long id) {
 
 				return objectHelperProvider
 					.find (id);
